+++ src/App.jsx
@@ -1,71 +1,22 @@
-import React, { useState } from "react";
-import rock from "./sources/rock.png";
-import paper from "./sources/paper.png";
-import scissors from "./sources/scissors.png";
+import React from "react";
+import Game from "./components/Game";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
+import Rules from "./components/Rules";
+import { Routes, Route } from "react-router-dom";
 
-const Button = (props) => {
+const App = () => {
   return (
-    <div value={props.name} onClick={props.onClick}>
-      <img class="user-selection-img" src={props.img} alt="img" />
-    </div>
-  );
-};
-function App() {
-  const [game, setGame] = useState({
-  name: "Player",
-  computer: "",
-  userSelection: "",
-  computerSelection: "",
-  message: "",
-});
-
-const playGame = (event) => {
-  const user = event.target.parentNode.getAttribute("value")
-    const computer = ["Rock", "Paper", "Scissors"][Math.floor(Math.random() * 3)];
-
-    if (user === computer) {
-      setGame({
-      ...game,
-      message: (game.message = "It's a tie!"),
-      });
-    } else if (
-      (user === "Rock" && computer === "Scissors") ||
-      (user === "Paper" && computer === "Rock") ||
-      (user === "Scissors" && computer === "Paper")
-    ) {
-      setGame({
-      ...game,
-      message: (game.message = "You won!"),
-      });
-    } else {
-      setGame({
-      ...game,
-      message: (game.message = "You lost!"),
-    });
-  }
-
-  setGame({
-    ...game,
-    userSelection: user,
-    computerSelection: computer,
-   });
-};
-
-  return (
-  <>
-    <h1>Rock, paper and scissors Game</h1>
-      <div class='ui buttons'>
-      <Button name="Rock" onClick={ playGame } img={rock} />
-      <Button name="Paper" onClick={ playGame } img={paper} />
-      <Button name="Scissors" onClick={ playGame } img={scissors}/>
-    </div>
-    <h1>{game.userSelection === "" ? "Pick one!" : `Your choice: ${game.userSelection}`}</h1>
-     <img className="pc-selection-img" src={ game.computerSelection === "Rock" ? rock : game.computerSelection === "Paper" ? paper : scissors } alt="img"/>
-      <h1>Computer choise: {game.computerSelection}</h1>
-      <h1 className="message">{game.message}</h1>
+    <>
+      <Header />
+      <Routes>
+        <Route path="/" element={<Game />} />
+        <Route path="/rules" element={<Rules />} />
+        <Game />
+      </Routes>
+      <Footer />
     </>
   );
-}
-
+};
 
 export default App;
