+++ src/components/Footer.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Footer = () => {
+  return (
+    <footer>
+      <p id="footer">Made with React {React.version}</p>
+    </footer>
+  );
+};
+
+export default Footer;
